+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add GvG tab (showing sector army details)
 * Fix GB calculation when self (still unsafe) is followed by owner
+* Fix (auto) battle info for newest version of FoE
 
 # 0.3.2
 * Fix battle info for newest version of FoE